+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this project will be documented in this file.
 
+## Unreleased
+
+### Maintenance
+
+* Add system tests
+
 ## [1.0.6]
 
 ### Maintenance